cloud_run_service/requirements.txt
Flask==3.0.3
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
gunicorn==22.0.0
Pillow==10.3.0
psycopg2-binary==2.9.10
google-cloud-storage==2.17.0
